Project ACP sub-agent transcripts into chat
