+++ src/client/features/chat/index.ts
@@ -10,6 +10,7 @@ export type { ChatInputProps } from './chat-input/chat-input';
 export { ChatInput } from './chat-input/chat-input';
 export { CompactingIndicator } from './compacting-indicator';
 export { PermissionPrompt, PermissionPromptExpanded } from './permission-prompt';
+export { projectAcpTranscriptUpdates } from './project-acp-transcript';
 export { QuestionPrompt } from './question-prompt';
 export { QueuedMessages } from './queued-messages';
 export type { TaskNotification } from './reducer';
