+++ src/shared/acp-protocol/index.ts
@@ -9,4 +9,5 @@ export * from './protocol/index';
 // Optional aliases for clearer ACP-oriented call sites.
 export type { ChatMessage as AgentChatMessage } from './protocol/messages';
 export type { SessionDeltaEvent as AgentSessionDeltaEvent } from './protocol/websocket';
+export * from './session-update-translator';
 export * from './subagents';
